// src/types/CountryType.ts
export type Country = {
  id: string;
  isoCode: string;
  title: string;
  name: string;
  img: string;
  overview: string;
  region: string;
  latitude: number;
  longitude: number;
  area: number;
  timeZone: string;
  language: string;
  population: number;
  currency: string;
  timeToTravel: string;
}